Move addAfterBranchSelector  to its own file
