+++ source/features/list-pr-for-branch.tsx
@@ -5,7 +5,7 @@ import {GitMergeIcon, GitPullRequestIcon} from '@primer/octicons-react';
 
 import features from '.';
 import getDefaultBranch from '../github-helpers/get-default-branch';
-import {addAfterBranchSelector} from './latest-tag-button';
+import addAfterBranchSelector from '../helpers/add-after-branch-selector';
 import {getPullRequestsAssociatedWithBranch} from './show-associated-branch-prs-on-fork';
 
 // Taken from https://github.com/fregante/github-issue-link-status/blob/98792f2837352bacbf80664f3edbcec8e579ed17/source/github-issue-link-status.js#L10
